app/requirements.txt
Flask
mysql-connector
flask-mysqldb
Flask-Session
Flask-WTF
flask-sitemapper
Flask-Bootstrap